[Ref] Silence signed/unsigned comparison warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@13168 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Vstplug.cpp
@@ -1043,7 +1043,7 @@ intptr_t CVstPlugin::Dispatch(VstOpcodeToPlugin opCode, int32 index, intptr_t va
 	if(exception)
 	{
 		mpt::ustring codeStr;
-		if(opCode < std::size(VstOpCodes))
+		if(opCode < static_cast<int32>(std::size(VstOpCodes)))
 			codeStr = mpt::ToUnicode(mpt::Charset::ASCII, VstOpCodes[opCode]);
 		else
 			codeStr = mpt::ufmt::val(opCode);
